requirements.txt
prettytable==3.17.0
python-dotenv==1.2.1
wcwidth==0.6.0
